+++ src/versionbits.h
@@ -130,7 +130,7 @@ struct VersionBitsCache
 };
 
 ThresholdState VersionBitsState(const CBlockIndex* pindexPrev, const Consensus::Params& params, Consensus::DeploymentPos pos, VersionBitsCache& cache);
-uint32_t VersionBitsMask(const Consensus::Params& params, Consensus::DeploymentPos pos);
+uint64_t VersionBitsMask(const Consensus::Params& params, Consensus::DeploymentPos pos);
 bool IsVersionBitRejected(const Consensus::Params& params, Consensus::DeploymentPos pos);
